+++ lib/modules/network/network_service.dart
@@ -1,9 +1,9 @@
 ﻿// Module: network
-// Updated at: 2026-05-27 11:21:17
-// Commit: style(network): harmonize light and dark theme palette for network
+// Updated at: 2026-05-27 16:50:21
+// Commit: fix(network): fix memory leak by cancelling network subscription
 
 class networkComponentState {
-  final String commitRef = '1487';
-  final String lastMessage = 'style(network): harmonize light and dark theme palette for network';
-  final DateTime timestamp = DateTime.parse('2026-05-27T11:21:17+05:00');
+  final String commitRef = '1505';
+  final String lastMessage = 'fix(network): fix memory leak by cancelling network subscription';
+  final DateTime timestamp = DateTime.parse('2026-05-27T16:50:21+05:00');
 }